Due to updating project file: fix: spaces rotor input and positionsright to left return most updated CodeObj
get notch from user:
codeObj now holds notches map
all c'tors updated and the use of tham
auto generated fixed
TODO- validateAndSetRotorsNotches
understand how to pars the string 4(2) (ex.) and fill notches map

toString, equals and hashCode to all classes

time calculation for process

+++ DTO/src/DTO/codeHistory/CodeHistory.java
@@ -15,7 +15,7 @@ public class CodeHistory {
         this.code = code;
     }
 
-    public void addTranslation(String input, String output, double time){
+    public void addTranslation(String input, String output, long time){
         inputOutput.add(new Translation(input, output, time));
     }
